requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
reportlab
google-generativeai